renamed to controller-utils.js
